+++ node/tests/constants.ts
@@ -20,6 +20,7 @@ export default {
     PANDOC_CONVERT_DIR_PATH: "/temp/convert/pandoc", // pandoc 转换目录
 
     BROADCAST_CHANNEL_NAME: "channel-test", // 测试用广播通道名称
+    BROADCAST_CHANNEL_NAME_MESSAGE: "channel-test", // 消息测试用广播通道名称
     /* 测试用文件内容 */
     TEST_FILE_CONTENT: `<html>
     <body>